Refactor path for NavigationMixin
