+++ src/backend/InvenTree/plugin/base/integration/NavigationMixin.py
